requirements.txt
streamlit>=1.30
opencv-python-headless>=4.9.0
numpy>=1.24
matplotlib>=3.7
Pillow>=9.5
scipy>=1.10